initial commit
The following code example demonstrates how to use the
javax.management.remote package to access an MBean server remotely.
Server is a server hosting an MBean server, while Client is a client
application to access the remote MBean server. In this example,
rmiregistry is used as a JNDI server. Remember to run rmiregistry first.